feat: Implemented a favorites section with the movies selected by the user.
The favorite feature is also updated on the heart displayed in the main
posters activity. If there are no favorite movies selected, a dialog will
be displayed and if the user unselectes all the previously favorite movies
the most popular section is displayed with a dialog indicating that there
are no favorite movies selected.